back_end/requirements.txt
asgiref==3.4.1
astroid==2.7.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.44.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
platformdirs==2.3.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
